requirements.txt
# 基础依赖
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0

# 训练相关
tqdm>=4.65.0

# 模型复杂度分析
thop>=0.1.1

# 数据处理
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.12.0

# 工具
pyyaml>=6.0
tensorboard>=2.12.0

# 可选：更快的数据加载
# opencv-python>=4.7.0
# albumentations>=1.3.0


